Classroom-Management-main/requirements.txt
Flask
numpy
pandas
matplotlib
flask_paginate
flask_sqlalchemy
pymysql
cryptography
mysql-connector-python
